backend/requirements.txt
fastapi>=0.115
uvicorn[standard]>=0.30
sqlalchemy[asyncio]>=2.0.25
asyncpg>=0.29
alembic>=1.13
pydantic>=2.6
pydantic-settings>=2.2
email-validator>=2.1
passlib[argon2]>=1.7.4
python-jose[cryptography]>=3.3.0
python-multipart>=0.0.9
httpx>=0.24.0
locust>=2.0.0
